+++ client/client_settings.hpp
@@ -1,6 +1,6 @@
 /***
     This file is part of snapcast
-    Copyright (C) 2014-2024  Johannes Pohl
+    Copyright (C) 2014-2025  Johannes Pohl
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -23,60 +23,101 @@
 #include "player/pcm_device.hpp"
 
 // standard headers
+#include <filesystem>
+#include <optional>
 #include <string>
 
 
-
+/// Snapclient settings
 struct ClientSettings
 {
+    /// Sharing mode for audio device
     enum class SharingMode
     {
-        unspecified,
-        exclusive,
-        shared
+        unspecified, ///< unspecified
+        exclusive,   ///< exclusice access
+        shared       ///< shared access
     };
 
+    /// Mixer settings
     struct Mixer
     {
+        /// Mixer mode
         enum class Mode
         {
-            hardware,
-            software,
-            script,
-            none
+            hardware, ///< hardware mixer
+            software, ///< software mixer
+            script,   ///< run a mixer script
+            none      ///< no mixer
         };
 
+        /// the configured mixer mode
         Mode mode{Mode::software};
-        std::string parameter{""};
+        /// mixer parameter
+        std::string parameter;
     };
 
+    /// Server settings
     struct Server
     {
-        std::string host{""};
+        /// server host or IP address
+        std::string host;
+        /// protocol: "tcp", "ws" or "wss"
+        std::string protocol{"tcp"};
+        /// server port
         size_t port{1704};
+        /// server certificate
+        std::optional<std::filesystem::path> server_certificate;
+        /// Certificate file
+        std::filesystem::path certificate;
+        /// Private key file
+        std::filesystem::path certificate_key;
+        /// Password for encrypted key file
+        std::string key_password;
+        /// Is ssl in use?
+        bool isSsl() const
+        {
+            return (protocol == "wss");
+        }
     };
 
+    /// The audio player (DAC)
     struct Player
     {
-        std::string player_name{""};
-        std::string parameter{""};
+        /// name of the player
+        std::string player_name;
+        /// player parameters
+        std::string parameter;
+        /// additional latency of the DAC [ms]
         int latency{0};
+        /// the DAC
         player::PcmDevice pcm_device;
+        /// Sampleformat to be uses, i.e. 48000:16:2
         SampleFormat sample_format;
+        /// The sharing mode
         SharingMode sharing_mode{SharingMode::unspecified};
+        /// Mixer settings
         Mixer mixer;
     };
 
+    /// Log settings
     struct Logging
     {
-        std::string sink{""};
+        /// The log sink (null,system,stdout,stderr,file:<filename>)
+        std::string sink;
+        /// Log filter
         std::string filter{"*:info"};
     };
 
+    /// The snapclient process instance
     size_t instance{1};
+    /// The host id, presented to the server
     std::string host_id;
 
+    /// Server settings
     Server server;
+    /// Player settings
     Player player;
+    /// Logging settings
     Logging logging;
 };