+++ common/message/client_info.hpp
@@ -1,6 +1,6 @@
 /***
     This file is part of snapcast
-    Copyright (C) 2014-2022  Johannes Pohl
+    Copyright (C) 2014-2025  Johannes Pohl
 
     This program is free software: you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
@@ -16,8 +16,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ***/
 
-#ifndef MESSAGE_CLIENT_INFO_HPP
-#define MESSAGE_CLIENT_INFO_HPP
+#pragma once
 
 // local headers
 #include "json_message.hpp"
@@ -27,39 +26,47 @@ namespace msg
 {
 
 /// Client information sent from client to server
-/// Might also be used for sync stats and latency estimations
+/// Might also be used for other things in future, like
+/// - sync stats
+/// - latency estimations
+/// - Battery status
+/// - ...
 class ClientInfo : public JsonMessage
 {
 public:
+    /// c'tor
     ClientInfo() : JsonMessage(message_type::kClientInfo)
     {
         setVolume(100);
         setMuted(false);
     }
 
+    /// d'tor
     ~ClientInfo() override = default;
 
+    /// @return the volume in percent
     uint16_t getVolume()
     {
         return get("volume", static_cast<uint16_t>(100));
     }
 
+    /// @return if muted or not
     bool isMuted()
     {
         return get("muted", false);
     }
 
+    /// Set the volume to @p volume percent
     void setVolume(uint16_t volume)
     {
         msg["volume"] = volume;
     }
 
+    /// Set muted to @p muted
     void setMuted(bool muted)
     {
         msg["muted"] = muted;
     }
 };
+
 } // namespace msg
-
-
-#endif